Move tests for CreateTempSubdirectory to separate file
